Status code fix
Returns the status code from proxied site.

This works and is compliant because, per RFC2616, “The first line of a
Response message is the Status-Line, consisting of the protocol version
followed by a numeric status code and its associated textual phrase,
with each element separated by SP characters…”

However, the HTTP-Version must be updated to whatever our server is on,
which is why only the Status-Code and Reason-Phrase of the proxied
server is returned.